requirements.txt
numpy==1.26.1
# pip==23.3.1
setuptools==69.0.0
six==1.16.0
wheel==0.41.3
opencv-python~=4.8.1.78
pyzbar~=0.1.9
python-dateutil==2.8.2
Pillow~=10.1.0
flask~=3.0.0
ipython~=8.17.2
treepoem~=3.23.0
pylibdmtx~=0.1.10
doctr~=1.9.0
requests~=2.31.0
pandas~=2.1.3
beautifulsoup4~=4.12.2
selenium~=4.15.2
weasyprint~=60.1 # (64-bit Windows 11 -> GTK3 -> environment variables -> Path -> Move Up)
# maybe this ? >>
python-doctr~=0.7.0
tf2onnx~=1.15.1
weasyprint~=60.1
email-validator~=2.1.0
tensorflow~=2.14.0
psycopg2==2.9.9
